requirements.txt
tensorflow==2.10
numpy == 1.24.3
Django == 4.2.13
djangorestframework == 3.15.2
opencv-python
mtcnn
keras-facenet
scikit-learn == 1.3.2
Pillow
drf-yasg
django-compressor
django-cors-headers
tensorflow-hub